server/pyproject.toml
[tool.poetry]
name = "ninsar server"
version = "0.1.0"
description = ""
authors = ["Kamil Biktimirov <Biktimirov.kamil@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
